Update travis config

+++ deploy.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+# vim: fileencoding=utf-8
 
 # CHANGE THE FOLLOWING VARIABLES
 GITHUB_REPO_URL="https://github.com/donizyo/LaTeX-Travis/"
@@ -7,10 +8,16 @@ GITHUB_REPO_URL="https://github.com/donizyo/LaTeX-Travis/"
 GITHUB_FULL_URL="https://donizyo:$GITHUB_TOKEN@${GITHUB_REPO_URL:8}"
 DEPLOY_BRANCH_NAME=deploy
 
-# Routines
-git checkout --orphan "$DEPLOY_BRANCH_NAME"
-git rm --cached --ignore-unmatch -rf -- .
-git add *.pdf
-git add *.log
-git commit --author="traviscibot <>" -m "Publish"
-git push "$GITHUB_FULL_URL" +"$DEPLOY_BRANCH_NAME" # Force push onto deploy branch
+if [ -z $1 ]; then
+  # argument list is needed by 'git add' in the following `else` block
+  echo "No argument is supplied!"
+  exit 1
+else
+  # Routines
+  git checkout --orphan "$DEPLOY_BRANCH_NAME"
+  git rm --cached --ignore-unmatch -rf -- .
+  git add -- "$@"
+  git commit --author="traviscibot <>" -m "Publish"
+  # Force push onto deploy branch
+  git push "$GITHUB_FULL_URL" +"$DEPLOY_BRANCH_NAME" 
+fi
